+++ apps/web/src/store/actions/siteBanners.js
@@ -17,6 +17,7 @@ const siteBannerFields = `
   type
   actionText
   actionUrl
+  showToNewUsers
   publishedAt
   unpublishedAt
   createdAt